# sort: implement -k and -t support