--- a/tests/by-util/test_sort.rs
+++ b/tests/by-util/test_sort.rs
@@ -398,6 +398,7 @@ aaaa
         .stdout_only(input);
 }
 
+#[test]
 fn test_zero_terminated() {
     test_helper("zero-terminated", "-z");
 }