For #1728.

This allows to specify keys after the -k flag and a custom field
separator using -t.

Not sure if I've missed something... For reference [GNU coreutils documentation about sort ](https://www.gnu.org/software/coreutils/manual/html_node/sort-invocation.html), [POSIX spec](https://pubs.opengroup.org/onlinepubs/9699919799/utilities/sort.html).