@electricboogie 2021-04-02
Just FYI, I've been working on sort too.  See most recently #2008.  Pleased to let you implement -k and -t, but be aware there is some functionality that I've added and I'd hate for us to needlessly conflict.  If you need my help with anything, please let me know.