@sylvestre 2021-04-02
```
---- test_sort::test_keys_invalid_field_zero stdout ----

current_directory_resolved: 

run: /home/travis/build/uutils/coreutils/target/debug/coreutils sort -k 0.1

thread 'test_sort::test_keys_invalid_field_zero' panicked at 'Broken pipe (os error 32)', tests/common/util.rs:742:37

note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace.
```

Is failing twice in the CI