@miDeb 2021-04-02
Well, that seems to have worked. There's one remaining failing test, saying
Error: Incompatible 'Cargo.lock' format; try `cargo +1.40.0 update`, which I guess isn't something I have caused because I haven't modified Cargo.lock at all.