@sylvestre 2021-04-02
I will fix it here:
https://github.com/uutils/coreutils/pull/2013