@miDeb 2021-04-10
This should be ready now. Performance seems to have regressed slightly, but I don't think it's a big deal. Note that if you want to measure performance you should specify an outfile (`-o`), because I made writes to stdout buffered as a drive-by change, resulting in a clear perf win for redirecting stdout (`sort > outfile`).