@sylvestre 2021-04-10
could you please create and document how you did the benchmarking?
For example, create a 
`BENCHMARKING.md `
file
with where to get a data set (or generate it) and the commands to evaluate performances? (with hyperfine if possible)

thanks