@miDeb 2021-04-10
I added a benchmarking.md file. Below is a comparison of performance before and after my change:

wordlist: (`hyperfine "target/release/coreutils sort shuffled_wordlist.txt -o output.txt"`)
before:
`Time (mean ± σ):      26.1 ms ±   3.6 ms    [User: 64.3 ms, System: 9.6 ms]`
after:
`Time (mean ± σ):      27.2 ms ±   2.3 ms    [User: 107.5 ms, System: 12.7 ms]`

numbers: (`hyperfine "target/release/coreutils sort shuffled_numbers.txt -n -o output.txt"`)
before:
`Time (mean ± σ):      40.1 ms ±   3.8 ms    [User: 338.7 ms, System: 9.7 ms]`
after:
`Time (mean ± σ):      44.6 ms ±   3.6 ms    [User: 370.2 ms, System: 9.5 ms]`